move json write and read utils to ftrack-utils and fix rosetta typo
